backend/requirements.txt
fastapi
uvicorn[standard]
pillow
python-dotenv
tensorflow
google-generativeai
python-multipart
pydantic
motor
pymongo
opencv-python-headless
